Use u64 behind the scenes for solana-tokens (#13815)
* Use u64 behind the scenes for Allocations

* Fixup readme

* Clippy and remove errant comments

+++ tokens/src/args.rs
@@ -9,11 +9,11 @@ pub struct DistributeTokensArgs {
     pub fee_payer: Box<dyn Signer>,
     pub stake_args: Option<StakeArgs>,
     pub spl_token_args: Option<SplTokenArgs>,
-    pub transfer_amount: Option<f64>,
+    pub transfer_amount: Option<u64>,
 }
 
 pub struct StakeArgs {
-    pub unlocked_sol: f64,
+    pub unlocked_sol: u64,
     pub stake_account_address: Pubkey,
     pub stake_authority: Box<dyn Signer>,
     pub withdraw_authority: Box<dyn Signer>,
